+++ apps/s3api/src/summary.rs
@@ -1,0 +1,12 @@
+use crate::S3ApiResult;
+
+pub type CommandResult = S3ApiResult<ResponseSummary>;
+
+#[derive(Debug)]
+pub struct ResponseSummary {}
+
+impl ResponseSummary {
+    pub fn empty() -> ResponseSummary {
+        ResponseSummary {}
+    }
+}
